--- plugins/tool/tool-version/api/tool-version.api
@@ -1,25 +1,0 @@
-public abstract interface class eu/bitfunk/gradle/plugin/tool/version/VersionContract {
-}
-
-public abstract interface class eu/bitfunk/gradle/plugin/tool/version/VersionContract$Generator {
-	public abstract fun generateFeatureVersionCode (Ljava/util/Date;)I
-	public abstract fun generateVersionCode ()I
-	public abstract fun generateVersionDetails ()Ljava/lang/String;
-	public abstract fun generateVersionName ()Ljava/lang/String;
-}
-
-public abstract interface class eu/bitfunk/gradle/plugin/tool/version/VersionContract$Plugin {
-	public abstract fun addPlugins (Lorg/gradle/api/Project;)V
-	public abstract fun configureVersion (Lorg/gradle/api/Project;Leu/bitfunk/gradle/plugin/tool/version/VersionContract$Generator;)V
-	public abstract fun configureVersionTasks (Lorg/gradle/api/Project;Leu/bitfunk/gradle/plugin/tool/version/VersionContract$Generator;)V
-}
-
-public final class eu/bitfunk/gradle/plugin/tool/version/VersionPlugin : eu/bitfunk/gradle/plugin/tool/version/VersionContract$Plugin, org/gradle/api/Plugin {
-	public fun <init> ()V
-	public fun addPlugins (Lorg/gradle/api/Project;)V
-	public synthetic fun apply (Ljava/lang/Object;)V
-	public fun apply (Lorg/gradle/api/Project;)V
-	public fun configureVersion (Lorg/gradle/api/Project;Leu/bitfunk/gradle/plugin/tool/version/VersionContract$Generator;)V
-	public fun configureVersionTasks (Lorg/gradle/api/Project;Leu/bitfunk/gradle/plugin/tool/version/VersionContract$Generator;)V
-}
-
